+++ src/types/wellknown.d.ts
@@ -1,4 +1,0 @@
-declare module "wellknown" {
-  export function parse(wkt: string): any;
-  export function stringify(geometry: any): string;
-}